Fix silly issues with date/time job information display...
Apologies, this one was totally on me, this was a bug that was patently obvious to anybody who didn't have my "developer's brain" tunnel vision when testing things.  Basically what happened was that I was needlessly parsing out the non-primary fields in the output from the atq command when they should've all just been displayed as one complete field group.